refactor: Refactor Select component using compound pattern for reusability
